Set Home zoning with fixed slots and flexible chart
Split HomeRootView into per-zone extension files. Fixed slots for
header, meal panel, and bottom controls (safe-area anchored); the
chart takes the remaining height. Drops safeAreaSize and the
adjustPadding device hacks; caps dashboard type size at XXL.
