+++ Trio/Sources/Modules/Home/View/Chart/ChartElements/CobIobChart.swift
@@ -44,7 +44,7 @@ extension MainChartView {
             "IOB": Color.darkerBlue
         ])
         .chartLegend(.hidden)
-        .frame(minHeight: geo.size.height * 0.12)
+        .frame(minHeight: cobIobHeight)
         .frame(width: fullWidth(viewWidth: screenSize.width))
         .chartXScale(domain: state.startMarker ... state.endMarker)
         .chartXSelection(value: $selection)
